Deck Selection and Users Hand
So basically I am currently creating 12 cards (currently just colored randomly) in the deck and holding them all in a stack. I made it so when you click on the deck a card is added to the users hand which is a list. Then I added a quick little hover animation type thing with a simple transformation on mouseOver. Gonna work on it a little more tn or tomorrow to add the ability to select a card and discard a card which should be pretty quick