Complete Milestone #1 (#21)
* Create my branch and scene with a platform and sphere

* Intentionally bad pull request

This is a bad pull request to preview the issues

* Revert "Intentionally bad pull request"

This reverts commit c1fc5cf02f1a75811ee32cc5b19a0c319111c059.

* Made testing scene

* moved camera down by 5 on y axis

* Big Cube

* Initial Scene Addition Test

Went ahead and added a scene w/ a cube

* Test2

* Erik Scene p.2

* Update gitignore

* Updated git ignore from dev branch

* weekend progress

* functional draw system prototype

* Add top bar of UI

* Deck Selection and Users Hand

So basically I am currently creating 12 cards (currently just colored randomly) in the deck and holding them all in a stack. I made it so when you click on the deck a card is added to the users hand which is a list. Then I added a quick little hover animation type thing with a simple transformation on mouseOver. Gonna work on it a little more tn or tomorrow to add the ability to select a card and discard a card which should be pretty quick

* Adjusted deck pos

* Restructure code and begin building main scene

* Delete .idea directory

* Clean up

* Delete .vsconfig

* turn system

* Finish UI Design & add targeting system

* Complete MileStone #1 and integrate in main scene

---------

Co-authored-by: Edward Leon <edleon761@gmail.com>
Co-authored-by: mbelles21 <mbelles@csumb.edu>
Co-authored-by: dconzemius <danieljconzmeius@gmail.com>
Co-authored-by: Erik <ermorris@csumb.edu>
Co-authored-by: Alexandre Pellereau <64319690+AlexandrePellereau@users.noreply.github.com>