Updated PlayerStats, Shop Fills with cards from CardDatabase and used their values.  Fixed bug in event scene and most of the scene should be working now.
Purchased cards are added to heldCard in the cardDatabase(I know this is changing but leaving it for now).  Will worry about relics and boosterPacks later
PlayerStats now has Coins-Glue-Tape variables
Event system should be pretty much working. I might need to make a video explaining how to make new events properly but I tried to comment the code thoroughly in the meantime.