Complete Milestone #3 (#50)
* Create my branch and scene with a platform and sphere

* Intentionally bad pull request

This is a bad pull request to preview the issues

* Revert "Intentionally bad pull request"

This reverts commit c1fc5cf02f1a75811ee32cc5b19a0c319111c059.

* Made testing scene

* moved camera down by 5 on y axis

* Big Cube

* Initial Scene Addition Test

Went ahead and added a scene w/ a cube

* Test2

* Erik Scene p.2

* Update gitignore

* Updated git ignore from dev branch

* weekend progress

* functional draw system prototype

* Add top bar of UI

* Deck Selection and Users Hand

So basically I am currently creating 12 cards (currently just colored randomly) in the deck and holding them all in a stack. I made it so when you click on the deck a card is added to the users hand which is a list. Then I added a quick little hover animation type thing with a simple transformation on mouseOver. Gonna work on it a little more tn or tomorrow to add the ability to select a card and discard a card which should be pretty quick

* Adjusted deck pos

* Restructure code and begin building main scene

* Delete .idea directory

* Clean up

* Delete .vsconfig

* Made card selections and turned "2D"

* Updated Card after merge conflict

* Card Fix

Accidentally removed a function in card but its back now

* turn system

* Finish UI Design & add targeting system

* Complete MileStone #1 and integrate in main scene

* Updated drawing system

* Functional map screen

* icon generation + map advancement

* Add new Deck system to main scene

* functional level select

* structure

* Added scripts via file upload

This is not a playable branch, as we need to figure out the deck drawing system (3D objects or UI). However the system is implemented, so in theory if we figure this issue out we can have a playable (This is not tested and there will likely be bugs)

* Made Shop and Event scenes

Also made 2 quick pngs, a resource folder and a shop.txt file. I will work more on the shop and event system later today.

* combatsystem

* Shop and Event essentially complete

Just need to have access to the card database and I need to know how the currency will work. Then just some small tweaks and they should be ready.

* Updated draw method, and integrate targeting and turn systems

* small changes

* Fix compile errors

* Updated code, still errors that will be fixed shortly

* Fix errors

* Fix errors

* Finish battle functionality

* Fix small bugs

* Fixing up card prefab

Set the blocks to public in the actual card prefab script so that we can link it up in the prefab view

* More tinkering w/ prefab

* Updated PlayerStats, Shop Fills with cards from CardDatabase and used their values.  Fixed bug in event scene and most of the scene should be working now.

Purchased cards are added to heldCard in the cardDatabase(I know this is changing but leaving it for now).  Will worry about relics and boosterPacks later
PlayerStats now has Coins-Glue-Tape variables
Event system should be pretty much working. I might need to make a video explaining how to make new events properly but I tried to comment the code thoroughly in the meantime.

* Add enemy logic to battle

* Restructure the scripts

* Game is rough but playable now

* New Assets for testing

Brought in some more basic assets for testing purposes, will add the block assets in a jif

* Clean Up && Additions

Added some more assets, and cleaned up the actual card script to organize the variables

* I hate the English Language

* Added final block art asset

I'm a comedic genius

* test

* Delete test.txt

* Quick fix

There was an error in regards to players not being able to select cards, that has now been fixed

* Basic implementation of updating visuals

Also updated the card datyabase as that's where the card will call two functions, whilst giving a cardId, to get the right name and image
***Images will be implemented in a hot minute***

* Shop Work/Rest Scene/Events/Scene Integration

Shop's now work with new cards (Had to add a case to ActualCard script.
Purchasing a card will also put the card into PlayerStats deck.
Made a quick rest scene.
Added more options for events.
Also made it so the map icons take you to the appropriate scene. Not the battle scenes though.

* enemy working

* fix player null

* renaming

* Fixed up visual stuff

The cards theoretically change appropriately depending on the card condition

* enemy fix and healthbars

* Update Battle Scene TestAlex.unity

* delete enemy type (useless)

* milestone #3 progress

* Add some file

* Scene integration, updated map icons, Player stats on ui updated

* Clean some stuff up

* Fixed shop with new update and rests/events

Health now works at rest sites and events
Noticed an error with PrePopulate being called multiple times in the database but have not fixed that entirely yet.
Fully working on relics now

* Change shop and event scripts to use game manager instance

* Fix last conflict

---------

Co-authored-by: Edward Leon <edleon761@gmail.com>
Co-authored-by: mbelles21 <mbelles@csumb.edu>
Co-authored-by: dconzemius <danieljconzmeius@gmail.com>
Co-authored-by: Erik <ermorris@csumb.edu>
Co-authored-by: Matthew Belles <142948584+mbelles21@users.noreply.github.com>
Co-authored-by: Alexandre Pellereau <64319690+AlexandrePellereau@users.noreply.github.com>
Co-authored-by: dconzemius <157443194+dconzemius@users.noreply.github.com>